% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{constructors}
\alias{constructors}
\alias{continuous}
\alias{binary}
\alias{categorical}
\title{Functions to define individual covariate properties}
\usage{
continuous(mean, sd)

binary(prob)

categorical(levels, probs = NULL)
}
\description{
These functions serve as constructors to create objects that define the
properties of different covariate types (continuous, binary, categorical).
These definition objects are then used within the \code{gen_data} function's
\code{covariate_defns} argument to specify the characteristics of the simulated covariates.
}
\details{
Constructors for Covariate Definitions
}
