% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{gen_data}
\alias{gen_data}
\title{Generate simulated datasets of IPD covariates and outcome for a trial}
\usage{
gen_data(
  N,
  b_0,
  b_trt,
  covariate_defns,
  b_prognostic = NULL,
  b_effect_modifier = NULL,
  cor_matrix = NULL,
  trt_assignment = list(prob_trt1 = 0.5),
  error_params = list(sd = 1),
  family = gaussian(link = "identity")
)
}
\arguments{
\item{cor_matrix}{Applies to all latent covariates}

\item{family}{}
}
\value{
Data frame of simulated IPD
}
\description{
Generate simulated datasets of IPD covariates and outcome for a trial
}
